Fixes mobile navbar
Moves the menu show/hide code to a JS controller:
  app/javascript/controllers/navbar.js

Added this attribute to the nav bar HTML:
data-controller="navbar"

This ensures the code in navbar.js is run automatically whenever the
nav bar is present. After adding that, everything just worked!
