+++ app/javascript/controllers/index.js
@@ -9,33 +9,3 @@ eagerLoadControllersFrom("controllers", application)
 // Lazy load controllers as they appear in the DOM (remember not to preload controllers in import map!)
 // import { lazyLoadControllersFrom } from "@hotwired/stimulus-loading"
 // lazyLoadControllersFrom("controllers", application)
-window.addEventListener("load", function () {
-  console.log(document.querySelector("#showMenu"));
-  document
-    .querySelector("#showMenu")
-    .addEventListener("click", function (event) {
-      document.querySelector("#mobileNav").classList.remove("hidden");
-    });
-
-  document
-    .querySelector("#hideMenu")
-    .addEventListener("click", function (event) {
-      document.querySelector("#mobileNav").classList.add("hidden");
-    });
-
-  document.querySelectorAll("[toggleElement]").forEach((toggle) => {
-    toggle.addEventListener("click", function (event) {
-      console.log(toggle);
-      const answerElement = toggle.querySelector("[answer]");
-      const caretElement = toggle.querySelector("img");
-      console.log(answerElement);
-      if (answerElement.classList.contains("hidden")) {
-        answerElement.classList.remove("hidden");
-        caretElement.classList.add("rotate-90");
-      } else {
-        answerElement.classList.add("hidden");
-        caretElement.classList.remove("rotate-90");
-      }
-    });
-  });
-});
