+++ app/views/shared/_navbar.html.erb
@@ -1,5 +1,5 @@
 <!-- Navigation -->
-<nav class="fixed flex justify-between w-full lg:px-48 md:px-12 px-4 content-center bg-secondary z-10">
+<nav data-controller="navbar" class="fixed flex justify-between w-full lg:px-48 md:px-12 px-4 content-center bg-secondary z-10">
   <div class="flex items-center">
     <%= link_to root_path do %>
       <%= image_tag 'AW_littlelogo.png', alt: "Logo", class: "h-40"%>
